pyproject.toml
[project]
name = "auralis"
version = "0.1.0"
description = "Auralis - the light that surrounds your documents."
authors = [
    {name = "Ted Ly",email = "lydongcanh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.118.3,<0.119.0)",
    "gql[all] (>=4.0.0,<5.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)"
]

[tool.poetry]
packages = [{include = "service_host", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
